+++ packages/react-dom/src/__tests__/ReactDOMServerIntegration-test.js
@@ -27,7 +27,6 @@ function initModules() {
   // First, reset the modules to load the client renderer.
   jest.resetModuleRegistry();
 
-  require('shared/ReactFeatureFlags').enableReactFragment = true;
   PropTypes = require('prop-types');
   React = require('react');
   ReactDOM = require('react-dom');
@@ -37,7 +36,6 @@ function initModules() {
   // Resetting is important because we want to avoid any shared state
   // influencing the tests.
   jest.resetModuleRegistry();
-  require('shared/ReactFeatureFlags').enableReactFragment = true;
   ReactDOMServer = require('react-dom/server');
 
   // Make them available to the helpers.
@@ -138,68 +136,6 @@ describe('ReactDOMServerIntegration', () => {
       expect(parent.childNodes[2].tagName).toBe('P');
     });
 
-    itRenders('a fragment with one child', async render => {
-      let e = await render(
-        <React.Fragment>
-          <div>text1</div>
-        </React.Fragment>,
-      );
-      let parent = e.parentNode;
-      expect(parent.childNodes[0].tagName).toBe('DIV');
-    });
-
-    itRenders('a fragment with several children', async render => {
-      let Header = props => {
-        return <p>header</p>;
-      };
-      let Footer = props => {
-        return (
-          <React.Fragment>
-            <h2>footer</h2>
-            <h3>about</h3>
-          </React.Fragment>
-        );
-      };
-      let e = await render(
-        <React.Fragment>
-          <div>text1</div>
-          <span>text2</span>
-          <Header />
-          <Footer />
-        </React.Fragment>,
-      );
-      let parent = e.parentNode;
-      expect(parent.childNodes[0].tagName).toBe('DIV');
-      expect(parent.childNodes[1].tagName).toBe('SPAN');
-      expect(parent.childNodes[2].tagName).toBe('P');
-      expect(parent.childNodes[3].tagName).toBe('H2');
-      expect(parent.childNodes[4].tagName).toBe('H3');
-    });
-
-    itRenders('a nested fragment', async render => {
-      let e = await render(
-        <React.Fragment>
-          <React.Fragment>
-            <div>text1</div>
-          </React.Fragment>
-          <span>text2</span>
-          <React.Fragment>
-            <React.Fragment>
-              <React.Fragment>
-                {null}
-                <p />
-              </React.Fragment>
-              {false}
-            </React.Fragment>
-          </React.Fragment>
-        </React.Fragment>,
-      );
-      let parent = e.parentNode;
-      expect(parent.childNodes[0].tagName).toBe('DIV');
-      expect(parent.childNodes[1].tagName).toBe('SPAN');
-      expect(parent.childNodes[2].tagName).toBe('P');
-    });
-
     itRenders('an iterable', async render => {
       const threeDivIterable = {
         '@@iterator': function() {
@@ -232,7 +168,6 @@ describe('ReactDOMServerIntegration', () => {
       // but server returns empty HTML. So we compare parent text.
       expect((await render(<div>{''}</div>)).textContent).toBe('');
 
-      expect(await render(<React.Fragment />)).toBe(null);
       expect(await render([])).toBe(null);
       expect(await render(false)).toBe(null);
       expect(await render(true)).toBe(null);
